Allow tagging subnets (all/public/private) via var.subnet_tags* (#19)
This adds two variables: `subnet_tags_public`, and `subnet_tags_private`. Each is a map with a `{}` default. If unset, current behavior is unchanged. If set to a map, the entries are merged into the final tags map. 

This is required to support alb-ingress-controller which performs [tag-based discovery](https://github.com/kubernetes-sigs/aws-alb-ingress-controller/blob/master/docs/setup.md#via-tags-on-the-subnets) to choose the subnets into which an ALB will be launched based on cluster/scheme.